Generalize TerrascriptCloudflareClientCollection to TerraformConfigClientCollection
